translation_agency/requirements.txt
google-adk==0.1.0
supabase>=2.13.0
litellm